requirements.txt
pyzmq
PySide2
PyOpenGL
